SortBy working.
One of the iterators is just a function that just takes an argument and
returns it.  If you sort by that iterator it does nothing but it kept
saying it would like it to be sorted. That's what it asked for so I made
it work despite being passed an argument that would make it specifically
not sort anything. Seems to me like it's not supposed to work this way
but it passes.